+++ config/typescript/tsconfig.runtime-testing.json
@@ -5,6 +5,8 @@
     "outDir": "../../.coverage-runtime-testing-temp"
   },
   "files": [
+    "../../runtime/testing/a11y-util.ts",
+    "../../runtime/testing/browser/index.ts",
     "../../runtime/testing/e2e/index.ts"
   ],
   "include": [
