+++ runtime/testing/e2e/a11y.ts
@@ -1,3 +1,5 @@
+import { SkyA11yUtil } from '../a11y-util';
+
 const logger = require('@blackbaud/skyux-logger');
 const axeBuilder = require('axe-webdriverjs');
 const protractor = require('protractor');
@@ -10,8 +12,6 @@ export abstract class SkyA11y {
       .then((url: string) => new Promise((resolve) => {
         const config = axeConfig.getConfig();
 
-        logger.info(`Starting accessibility checks for ${url}...`);
-
         axeBuilder(protractor.browser.driver)
           .options(config)
           .analyze((results: any) => {
@@ -21,31 +21,12 @@ export abstract class SkyA11y {
             logger.info(`Accessibility checks finished with ${numViolations} ${subject}.\n`);
 
             if (numViolations > 0) {
-              SkyA11y.logViolations(results);
+              const message = SkyA11yUtil.parseMessage(results.violations);
+              logger.error(message);
             }
 
             resolve(numViolations);
           });
       }));
-  }
-
-  private static logViolations(results: any): void {
-    results.violations.forEach((violation: any) => {
-      const wcagTags = violation.tags
-        .filter((tag: string) => tag.match(/wcag\d{3}|^best*/gi))
-        .join(', ');
-
-      const html = violation.nodes.reduce((accumulator: string, node: any) => {
-        return `${accumulator}\n${node.html}\n`;
-      }, '       Elements:\n');
-
-      const error = [
-        `aXe - [Rule: \'${violation.id}\'] ${violation.help} - WCAG: ${wcagTags}`,
-        `       Get help at: ${violation.helpUrl}\n`,
-        `${html}\n\n`
-      ].join('\n');
-
-      logger.error(error);
-    });
   }
 }
